Enable S2S and upload; add proper comparison
